Rename prometheus library same as interface name
A recent convention has been adopted to ensure charm library
names have the same name as the interface the service. This
commit ensures that Prometheus charm library confirms with
this convention.
