+++ tests/test_endpoint_provider.py
@@ -6,7 +6,7 @@ import re
 import unittest
 from unittest.mock import patch
 
-from charms.prometheus_k8s.v0.prometheus import MetricsEndpointProvider
+from charms.prometheus_k8s.v0.prometheus_scrape import MetricsEndpointProvider
 from ops.charm import CharmBase
 from ops.framework import StoredState
 from ops.testing import Harness
